// tools/conformance/src/modules/nestjs-service-module-generator/nestjs-service-module-generator.types.ts
// 🏷️ Types

import type { Tree } from "@nx/devkit";

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsServiceModuleGeneratorArguments {
  options: NestjsServiceModuleGeneratorOptions;
  tree: Tree;
}

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsServiceModuleGeneratorOptions {
  name: string;
  project?: string;
}